Ticket 0035: TLC — rename common.py to result_row.py, remove unused import
Rename the junk-drawer matching/common.py to result_row.py (describes its
single responsibility: building reconciliation result dicts). Update all
imports in lp.py, phased.py, and README.md. Remove unused Path import from
test_matching_score.py flagged by ruff.
