+++ src/aedist/matching/result_row.py
